requirements.txt
pytorch
pandas
scikit-image
matplotlib
torchvision
warnings